feat: Disconnected can now detect if you have that session running and will attach instead of creating more windows in your current session.
This was something that I ran into occasionally, I forgot to check and
see if I had already run disconnected and started a specific session. So
when I would start it again, it would relaunch all the windows in that
session and just make things a royal mess. So I have attached the shell
command for tmux ls, and I search that output to see if the name of the
session is found in there. It is not a perfect solution, I would like to
parse the output better than how I have it, but it will work for now.
Naming any session something like window will cause weird behavior.